Split visor rpc_client.go (2061 lines) into real and mock clients
- rpc_client.go (1023 lines): real RPC client implementation
- rpc_client_mock.go (1066 lines): mock client for testing
